feat: add 'publications and talks' section
This PR implements "Publications & Talks" section. This section must have a set of videos of publications and/or talks and a small description.
For that reason, I've created a `js` file that exports an array of objects. Each object represents one publication/talk and it must have:

- link - `string` - URL to the corresponding video
- description - `string` - ID of `react-intl` translation that holds talk/publication short description
- active - `boolean`- just one of the objects must have this property as `true`. It indicates which video should be highlighted when the page loads.

Thus, if one wants to add a new publication/talk, one just need to: 

1. add a short description on each file of `react-intl` messages; 

2. add a new object to the `array` on [this](https://github.com/ipfs/js.ipfs.io/blob/feat/publications-and-talks-section/src/shared/data/publications-and-talks/index.js) file.