fix: improvements based on feedback
This PR improves:

- David's feedback - **Hero section**

    1. Alpha banner (disclaimer)
    2. Urls centered (navbar)
    3. Hero title is not uppercase anymore
    4. Hero description is narrower
    5. Npm info is bigger (margin between info and font-size)
    6. Npm info is automatic now, using [npms api](https://api-docs.npms.io/) (see https://github.com/ipfs/js.ipfs.io/issues/21)
    7. Spinning `SVG` lines around the cube
    8. Cube size

- Previous improvements: 

    1. fix font sizes for mobile devices and added `Roboto` font
    2. added automatic scroll to the highlighted video when a user clicks on a new video
    3. fix **features section** media blocks margins

We are using [`date-fns`](https://github.com/date-fns/date-fns) to get the distance between dates in words. For that reason, we need to handle i18n. `date-fns` handle it for us, but we need to import the right locale. However, we do not need to import all the `date-fns` [supported locales](https://date-fns.org/v1.29.0/docs/I18n#supported-languages). We just need to import the locales that we are supporting on `js.ipfs.io`. So, if one needs to add a new locale to `js.ipfs.io`, one also needs to update [`dateFnsLocales.js`](https://github.com/ipfs/js.ipfs.io/blob/fix/improvements/src/utils/dateFnsLocales.js) file. 

This info will also be on the final `readme`.