chore: add service worker tests (#211)
This PR adds end-to-end (E2E) testing to `service worker gateway` feature, using [cypress](https://github.com/cypress-io/cypress). Shortly, the main goal of this test is to check whether the service worker is getting registered/unregistered properly or not when triggered. 

This test makes sure that we are not deploying the website with issues on `service-worker-gateway` feature.

**NOTE:** Currently, we've created a `pre-push` hook where the test runs. This is just a temporary solution as the main goal is to have a post-build stage on the CI to run the test.

This PR is related to #199.