chore: update ipfs dep (#267)
This PR updates ipfs to 0.37.0-rc.0.

Due to the constructor changes in this version the examples have also been updated.

This required an update of `react-live` to a version that supports async/await. Unfortunately in this new version styling for syntax highlighting is done by "theme" objects not CSS so there are a bunch of changes that are fallout from this.

Notably, I've updated the `syntax-highlighter` component to use the same highlight method as `react-live` so that we can re-use the theme object and not also import the theme as a CSS file.

I also switchied to using the `Prism` instance provided by `prism-react-renderer` that `react-live` uses instead of importing `prismjs`. This should reduce overall bundle size.

License: MIT
Signed-off-by: Alan Shaw <alan.shaw@protocol.ai>